Check if userControlledMassesUpdater is a function. See https://github.com/phetsims/balancing-act/issues/158.

+++ js/common/model/Mass.js
@@ -86,9 +86,11 @@ class Mass extends PhetioObject {
 
     // Must relink and unlink userControlledMassesUpdater after all other listeners to correctly update
     // userControlledMasses and activeDropPositions due to listener order dependency.
-    this.userControlledProperty.link( this.userControlledMassesUpdater );
-    this.userControlledProperty.reset();
-    this.userControlledProperty.unlink( this.userControlledMassesUpdater );
+    if ( this.userControlledMassesUpdater ) {
+      this.userControlledProperty.link( this.userControlledMassesUpdater );
+      this.userControlledProperty.reset();
+      this.userControlledProperty.unlink( this.userControlledMassesUpdater );
+    }
     super.dispose();
   }
